+++ src/cache/core/types/Cache.ts
@@ -1,4 +1,5 @@
 import { DataProxy } from './DataProxy';
+import { Modifier, Modifiers } from './common';
 
 export namespace Cache {
   export type WatchCallback = (diff: Cache.DiffResult<any>) => void;
@@ -27,9 +28,16 @@ export namespace Cache {
   }
 
   export interface EvictOptions {
-    id: string;
+    id?: string;
     fieldName?: string;
     args?: Record<string, any>;
+    broadcast?: boolean;
+  }
+
+  export interface ModifyOptions {
+    id?: string;
+    fields: Modifiers | Modifier<any>;
+    optimistic?: boolean;
     broadcast?: boolean;
   }