+++ src/react/context/ApolloContext.ts
@@ -1,5 +1,5 @@
+import React from 'react';
 import { ApolloClient } from '../../ApolloClient';
-import { requireReactLazily } from '../react';
 
 export interface ApolloContextValue {
   client?: ApolloClient<object>;
@@ -20,7 +20,6 @@ const contextSymbol = typeof Symbol === 'function' && Symbol.for ?
   '__APOLLO_CONTEXT__';
 
 export function resetApolloContext() {
-  const React = requireReactLazily();
   Object.defineProperty(React, contextSymbol, {
     value: React.createContext<ApolloContextValue>({}),
     enumerable: false,
@@ -30,7 +29,6 @@ export function resetApolloContext() {
 }
 
 export function getApolloContext() {
-  const React = requireReactLazily();
   if (!(React as any)[contextSymbol]) {
     resetApolloContext();
   }